+++ src/store/explorer.ts
@@ -1,7 +1,7 @@
 import { atom } from "jotai";
 import { NoteInfo, Obsidian } from "@/models";
 import { dataDir } from "@tauri-apps/api/path";
-import { CONFIG_FILE_NAME, readDirectory, readFile, writeFile } from "@/utils";
+import { CONFIG_FILE_NAME, readDirectory, readFile, writeFile } from "@/libs";
 
 const dataDirPath = (await dataDir()) + CONFIG_FILE_NAME;
 const openedFolderPath = atom<string>("");
